requirements.txt
ollama==0.1.0 
transformers==4.30.2 
sentence-transformers==2.3.1 
langchain==0.1.0 
python-dotenv==1.0.0 
scikit-learn==1.3.2
PyPDF2==3.0.1
python-docx==0.8.11
pandas==2.0.1
flask==2.3.2
streamlit>=1.31.0
plotly==5.15.0
numpy==1.24.3
scipy==1.10.1
fairlearn==0.8.0
openai==1.3.7